+++ src/android/src/main/java/com/ialocation/plugin/IALocationPlugin.java
@@ -24,6 +24,7 @@ import com.indooratlas.android.sdk.IAOrientationListener;
 import com.indooratlas.android.sdk.IARadioScanRequest;
 import com.indooratlas.android.sdk.IAWayfindingListener;
 import com.indooratlas.android.sdk.IAWayfindingRequest;
+import com.indooratlas.android.sdk.IAWayfindingTags;
 import com.indooratlas.android.sdk.IAGeofence;
 import com.indooratlas.android.sdk.IAGeofenceRequest;
 import com.indooratlas.android.sdk.resources.IALatLngFloor;
@@ -65,9 +66,11 @@ public class IALocationPlugin extends CordovaPlugin {
     private CallbackContext mCbContext;
     private IndoorLocationListener mListener;
     private boolean mLocationServiceRunning = false;
+    private boolean mDestroyed = false;
     private IALocationRequest mLocationRequest = IALocationRequest.create();
-    private IAOrientationRequest mOrientationRequest = new IAOrientationRequest(1.0, 1.0);
+    private IAOrientationRequest mOrientationRequest = new IAOrientationRequest(-1.0, 10.0);
     private IARadioScanRequest mRadioScanRequest = null;
+    private String mApiKey;
 
     /**
      * Called by the WebView implementation to check for geolocation permissions, can be used
@@ -136,7 +139,12 @@ public class IALocationPlugin extends CordovaPlugin {
     public boolean execute(String action, JSONArray args, com.remobile.cordova.CallbackContext realCallbackContext) throws JSONException {
         // react.native
         CallbackContext callbackContext = new CallbackContext(action, realCallbackContext, this);
-        try{
+        if (mDestroyed) {
+            Log.w(TAG, "already destroyed, ignoring action " + action);
+            callbackContext.error(PositionError.getErrorObject(PositionError.ALREADY_DESTROYED));
+            return false;
+        }
+        try {
             if ("initializeIndoorAtlas".equals(action)) {
                 if (validateIAKeys(args)) {
                     String apiKey = args.getString(0);
@@ -202,35 +210,24 @@ public class IALocationPlugin extends CordovaPlugin {
               addAttitudeCallback(callbackContext);
             } else if ("removeAttitudeCallback".equals(action)) {
               removeAttitudeCallback();
-            } else if ("addHeadingCallback".equals(action)) {
-              addHeadingCallback(callbackContext);
-            } else if ("removeHeadingCallback".equals(action)) {
-              removeHeadingCallback();
             } else if ("setSensitivities".equals(action)) {
               double orientationSensitivity = args.getDouble(0);
-              double headingSensitivity = args.getDouble(1);
-              setSensitivities(orientationSensitivity, headingSensitivity, callbackContext);
+              setSensitivities(orientationSensitivity, callbackContext);
             } else if ("addStatusChangedCallback".equals(action)) {
               addStatusChangedCallback(callbackContext);
             } else if ("removeStatusCallback".equals(action)) {
               removeStatusCallback();
             } else if ("requestWayfindingUpdates".equals(action)) {
-              double lat = args.getDouble(0);
-              double lon = args.getDouble(1);
-              int floor = args.getInt(2);
-              requestWayfindingUpdates(lat, lon, floor, callbackContext);
+              IAWayfindingRequest req = getWayfindingRequestFromJSON(args.getJSONObject(0));
+              requestWayfindingUpdates(req, callbackContext);
             } else if ("requestWayfindingRoute".equals(action)) {
               JSONObject _from = args.getJSONObject(0);
-              JSONObject _to = args.getJSONObject(1);
               IALatLngFloor from = new IALatLngFloor(
                   _from.getDouble("latitude"),
                   _from.getDouble("longitude"),
                   _from.getInt("floor"));
-              IALatLngFloor to = new IALatLngFloor(
-                  _to.getDouble("latitude"),
-                  _to.getDouble("longitude"),
-                  _to.getInt("floor"));
-              requestWayfindingRoute(from, to, callbackContext);
+              IAWayfindingRequest req = getWayfindingRequestFromJSON(args.getJSONObject(1));
+              requestWayfindingRoute(from, req, callbackContext);
             } else if ("removeWayfindingUpdates".equals(action)) {
               removeWayfindingUpdates();
             } else  if ("watchGeofences".equals(action)) {
@@ -261,9 +258,9 @@ public class IALocationPlugin extends CordovaPlugin {
               clearWifiWatch();
             }
         }
-        catch(Exception ex) {
-            Log.e(TAG,ex.toString());
-            callbackContext.error(PositionError.getErrorObject(PositionError.UNSPECIFIED_ERROR,ex.toString()));
+        catch (Exception ex) {
+            Log.e(TAG, ex.getMessage(), ex);
+            callbackContext.error(PositionError.getErrorObject(PositionError.UNSPECIFIED_ERROR,ex.getMessage()));
             return false;
         }
         return true;
@@ -276,6 +273,9 @@ public class IALocationPlugin extends CordovaPlugin {
     public void onDestroy() {
         if (mLocationManager != null){
             mLocationManager.destroy();
+            mLocationManager = null;
+            mLocationServiceRunning = false;
+            mDestroyed = true;
         }
         super.onDestroy();
     }
@@ -306,7 +306,8 @@ public class IALocationPlugin extends CordovaPlugin {
      * @param apiSecret
      */
     private void initializeIndoorAtlas(final String apiKey, final String apiSecret, final String pluginVersion, final CallbackContext callbackContext) {
-        if (mLocationManager == null){
+        if (mLocationManager == null || !apiKey.equals(mApiKey)) {
+            mApiKey = apiKey;
             cordova.getActivity().runOnUiThread(new Runnable() {
                 @Override
                 public void run() {
@@ -318,10 +319,15 @@ public class IALocationPlugin extends CordovaPlugin {
                     if (pluginVersion != null) {
                         bundle.putString("com.indooratlas.android.sdk.intent.extras.wrapperVersion", pluginVersion);
                     }
+                    if (mLocationManager != null) {
+                        mLocationManager.destroy();
+                    }
                     mLocationManager = IALocationManager.create(cordova.getActivity().getApplicationContext(), bundle);
                     callbackContext.success();
                 }
             });
+        } else {
+            callbackContext.success();
         }
     }
 
@@ -346,13 +352,6 @@ public class IALocationPlugin extends CordovaPlugin {
      */
     private void addAttitudeCallback(CallbackContext callbackContext) {
       getListener(this).addAttitudeCallback(callbackContext);
-    }
-
-    /**
-     * Adds a new callback to the IndoorAtlas IAAttitude.Listener
-     */
-    private void addHeadingCallback(CallbackContext callbackContext) {
-      getListener(this).addHeadingCallback(callbackContext);
     }
 
     /**
@@ -385,13 +384,6 @@ public class IALocationPlugin extends CordovaPlugin {
       getListener(this).removeAttitudeCallback();
     }
 
-    /**
-     * Removes callback from IndoorAtlas location listener
-     */
-     private void removeHeadingCallback() {
-       getListener(this).removeHeadingCallback();
-     }
-
      /**
       * Removes callback from IndoorAtlas location listener
       */
@@ -405,18 +397,12 @@ public class IALocationPlugin extends CordovaPlugin {
      * 2) Set destination of the wayfinder instance
      * 3) Get route between the given location and destination
      */
-    private void requestWayfindingUpdates(Double lat, Double lon, int floor, CallbackContext callbackContext) {
+    private void requestWayfindingUpdates(final IAWayfindingRequest req, CallbackContext callbackContext) {
         getListener(this).requestWayfindingUpdates(callbackContext);
-        final IAWayfindingRequest wayfindingRequest = new IAWayfindingRequest.Builder()
-            .withLatitude(lat)
-            .withLongitude(lon)
-            .withFloor(floor)
-            .build();
-
         cordova.getActivity().runOnUiThread(new Runnable() {
             @Override
             public void run() {
-              mLocationManager.requestWayfindingUpdates(wayfindingRequest, getListener(IALocationPlugin.this));
+              mLocationManager.requestWayfindingUpdates(req, getListener(IALocationPlugin.this));
             }
         });
     }
@@ -433,7 +419,7 @@ public class IALocationPlugin extends CordovaPlugin {
 
     private void requestWayfindingRoute(
         final IALatLngFloorCompatible from,
-        final IALatLngFloorCompatible to,
+        final IAWayfindingRequest to,
         final CallbackContext callbackContext
     ) {
         final IAWayfindingListener listener = new IAWayfindingListener() {
@@ -568,8 +554,8 @@ public class IALocationPlugin extends CordovaPlugin {
     /**
      * Set sensitivities for orientation and heading filter
      */
-     private void setSensitivities(double orientationSensitivity, double headingSensitivity, CallbackContext callbackContext) {
-       mOrientationRequest = new IAOrientationRequest(headingSensitivity, orientationSensitivity);
+     private void setSensitivities(double orientationSensitivity, CallbackContext callbackContext) {
+       mOrientationRequest = new IAOrientationRequest(-1.0, orientationSensitivity);
        cordova.getActivity().runOnUiThread(new Runnable() {
            @Override
            public void run() {
@@ -595,11 +581,7 @@ public class IALocationPlugin extends CordovaPlugin {
      */
     private boolean validateIAKeys(JSONArray args) throws JSONException {
         String apiKey = args.getString(0);
-        String apiSecret = args.getString(1);
         if (apiKey.trim().equalsIgnoreCase("")) {
-            return false;
-        }
-        if (apiSecret.trim().equalsIgnoreCase("")) {
             return false;
         }
         return true;
@@ -1043,5 +1025,47 @@ public class IALocationPlugin extends CordovaPlugin {
     @ReactMethod
     public void clearWifiWatch(ReadableArray args, Callback success, Callback error) {
         executeReactMethod("clearWifiWatch", args, success, error);
+    }
+
+    // Required for rn built in EventEmitter Calls.
+    @ReactMethod
+    public void addListener(String eventName) {
+    }
+
+    @ReactMethod
+    public void removeListeners(Integer count) {
+    }
+
+    private static IAWayfindingRequest getWayfindingRequestFromJSON(JSONObject json) throws JSONException {
+        IALatLngFloor to = new IALatLngFloor(
+            json.getDouble("latitude"),
+            json.getDouble("longitude"),
+            json.getInt("floor")
+        );
+        IAWayfindingRequest.Builder req = new IAWayfindingRequest.Builder().withDestination(to);
+        if (json.has("tags")) {
+            JSONObject tags = json.getJSONObject("tags");
+            IAWayfindingTags.Mode includeMode = "all".equals(tags.getString("includeMode")) ?
+                IAWayfindingTags.Mode.ALL : IAWayfindingTags.Mode.ANY;
+            IAWayfindingTags.Mode excludeMode = "all".equals(tags.getString("excludeMode")) ?
+                IAWayfindingTags.Mode.ALL : IAWayfindingTags.Mode.ANY;
+            List<String> includeTags = new ArrayList<>();
+            JSONArray includeTags_ = tags.getJSONArray("includeTags");
+            for (int i = 0; i < includeTags_.length(); i++) {
+                includeTags.add(includeTags_.getString(i));
+            }
+            List<String> excludeTags = new ArrayList<>();
+            JSONArray excludeTags_ = tags.getJSONArray("excludeTags");
+            for (int i = 0; i < excludeTags_.length(); i++) {
+                excludeTags.add(excludeTags_.getString(i));
+            }
+            req.withTags(new IAWayfindingTags(
+                includeTags,
+                excludeTags,
+                includeMode,
+                excludeMode
+            ));
+        }
+        return req.build();
     }
 }